backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
mysql-connector-python==8.2.0
passlib==1.7.4
google-generativeai
python-dotenv==1.0.0
werkzeug==3.0.1
PyPDF2==3.0.1
Pillow==10.1.0
python-docx